requirements.txt
openai==0.27.0
ratelimiter==1.2.0.post0
retrying==1.3.4
tiktoken==0.3.0
fluidsynth 
pyfluidsynth
music21
python-dotenv
pretty_midi
scipy